+++ v8go.cc
@@ -835,6 +835,23 @@ ValuePtr NewValueIntegerFromUnsigned(IsolatePtr iso, uint32_t v) {
   return tracked_value(ctx, val);
 }
 
+RtnValue NewStringFromBytes(IsolatePtr iso, const uint8_t* v, int len){
+  ISOLATE_SCOPE_INTERNAL_CONTEXT(iso);
+  TryCatch try_catch(iso);
+  RtnValue rtn = {};
+  Local<String> str;
+  if (!String::NewFromOneByte(iso, v, NewStringType::kNormal, len).ToLocal(&str)) {
+    rtn.error = ExceptionError(try_catch, iso, ctx->ptr.Get(iso));
+    return rtn;
+  }
+  m_value* val = new m_value;
+  val->iso = iso;
+  val->ctx = ctx;
+  val->ptr = Persistent<Value, CopyablePersistentTraits<Value>>(iso, str);
+  rtn.value = tracked_value(ctx, val);
+  return rtn; 
+}
+
 RtnValue NewValueString(IsolatePtr iso, const char* v, int v_length) {
   ISOLATE_SCOPE_INTERNAL_CONTEXT(iso);
   TryCatch try_catch(iso);
@@ -997,6 +1014,7 @@ RtnString ValueToString(ValuePtr ptr) {
   // TODO: Consider propagating the JS error. A fallback value could be returned
   // in Value.String()
   String::Utf8Value src(iso, value);
+  printf("___ valuetostring %d \n", src.length());
   char* data = static_cast<char*>(malloc(src.length()));
   memcpy(data, *src, src.length());
   rtn.data = data;